Refactor and pdf documen tree reader pdf syntax tree from render logic. Added PDF annotation and outline readers.
