+++ src/renderer/annotation/AnnotationHelpers.js
