+++ src/renderer/file-browser/browserApp.js
@@ -1,6 +1,6 @@
 import React, { useState, useReducer } from "react";
 import { getCurrentPath, handleFileSelect, handleFolderSelect } from "@/actions";
-import { listDir, watchOrReload } from "@/fileHelpers";
+import { listDir, watchOrReload } from "@/file-browser/fileHelpers";
 import "file-icon-vectors/dist/file-icon-square-o.css";
 import "@/css/browser.scss";
