+++ src/renderer/file-browser/fileHelpers.js
